move macos icons to initializers
