+++ src/lib/settings/initializers.ts
@@ -1,13 +1,13 @@
+import {frameUrls, iconUrls} from "$lib/utils/macicon";
 import {fetchThemeFiles, parseThemeFiles} from "$lib/utils/themes";
-import {registry} from "./registry";
-import type {SettingsRegistry} from "./types";
+import {registry, type SettingSchema} from "./registry";
 
 
-type Initializer = (registry: SettingsRegistry) => Promise<void>;
+type Initializer = (registry: SettingSchema) => Promise<void>;
 
 const initializers: Initializer[] = [
     // Theme list from iTerm2-Color-Schemes
-    async (reg: SettingsRegistry) => {
+    async (reg: SettingSchema) => {
         const themeSetting = reg.theme;
         if (themeSetting?.type !== "theme") return;
         const files = await fetchThemeFiles();
@@ -15,8 +15,35 @@ const initializers: Initializer[] = [
         themeSetting.options.push(...themeNames);
     },
 
+    // eslint-disable-next-line @typescript-eslint/require-await
+    async (reg: SettingSchema) => {
+        const iconSetting = reg.macosIcon;
+        if (iconSetting?.type !== "dropdown") return;
+        iconSetting.options = [
+            ...Object.keys(iconUrls).map(key => ({
+                value: key,
+                name: key[0].toUpperCase() + key.slice(1),
+                group: "Predefined icons",
+                icon: iconUrls[key]
+            })),
+            {value: "custom", name: "Custom Icon", description: "Use your own icon file.", group: "Custom"},
+            {value: "custom-style", name: "Custom Style", description: "Customize the icon with colors and frames.", group: "Custom"}
+        ];
+    },
+
+    // eslint-disable-next-line @typescript-eslint/require-await
+    async (reg: SettingSchema) => {
+        const frameSetting = reg.macosIconFrame;
+        if (frameSetting?.type !== "dropdown") return;
+        frameSetting.options = Object.keys(frameUrls).map(key => ({
+            value: key,
+            name: key[0].toUpperCase() + key.slice(1),
+            icon: frameUrls[key]
+        }));
+    },
+
     // Font list... web: skip or no-op; desktop: shell out to `ghostty +list-fonts`
-    async (_: SettingsRegistry) => {
+    async (_: SettingSchema) => {
         // The following TODO was written by a clanker.
         // TODO: implement this, ideally by shelling out to `ghostty +list-fonts`
         // and parsing the output, which would ensure the list is accurate and
